Use my version of credit card creation
IMO the version in master is a bit smelly - you'd think this would belong in
credit_card.rb. This version also remembers to update the customer hash when
creating a credit card for a customer. And lastly, this runs a check against the
decline_all_cards/verify_all_cards flags, which should be useful for testing
credit card TR.