Hack to make fake_braintree work with poltergeist.
When using poltergeist/phantomjs there is an issue when the backend
capybara server makes the Braintree::TransparentRedirect.confirm call
across to the fake_braintree sinatra app running on another thread from
the same overall process. This causes the tests to hang and eventually
time out.

By forking the process this issue no longer occurs and the two processes
can communicate with each other.

I'm not too sure on how to properly deal with forked processes, as we
detach in order to continue execution it does leave a Thin server off on
it's own after the test run is complete. You can fix this by adding an
at_exit hook to your cucumber/rspec suite.

at_exit do
  Process.kill("QUIT", ENV["BRAINTREE_PROCESS"].to_i)
end